updating the api client for getting the responsible officer
