+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsinterventionsservice/component/RamDeliusClient.kt
@@ -4,6 +4,7 @@ import org.springframework.stereotype.Component
 
 @Component
 class RamDeliusClient(private val ramDeliusApiClient: RestClient) {
+
   fun makeSyncPutRequest(uri: String, requestBody: Any) {
     ramDeliusApiClient.put(uri, requestBody).retrieve().toBodilessEntity().block()
   }
