+++ src/test/kotlin/uk/gov/justice/digital/hmpps/hmppsinterventionsservice/service/CommunityAPIOffenderServiceTest.kt
@@ -3,17 +3,15 @@ package uk.gov.justice.digital.hmpps.hmppsinterventionsservice.service
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
-import org.mockito.kotlin.mock
 import org.springframework.http.HttpStatus
 import org.springframework.web.reactive.function.client.ClientResponse
 import org.springframework.web.reactive.function.client.WebClient
 import org.springframework.web.reactive.function.client.WebClientResponseException
 import reactor.core.publisher.Mono
 import uk.gov.justice.digital.hmpps.hmppsinterventionsservice.component.RestClient
-import uk.gov.justice.digital.hmpps.hmppsinterventionsservice.config.InvalidAssumptionError
 import uk.gov.justice.digital.hmpps.hmppsinterventionsservice.util.AuthUserFactory
 
-private data class MockedResponse(
+data class MockedResponse(
   val path: String,
   val status: HttpStatus,
   val responseBody: String,
@@ -26,8 +24,6 @@ internal class CommunityAPIOffenderServiceTest {
   private val staffDetailsLocation = "staff-details"
   private val offenderManagersLocation = "offender-managers"
   private val offenderIdentifiersLocation = "offender-identifiers"
-  private val responsibleOfficerLocation = "responsible-officer"
-  private val telemetryService = TelemetryService(mock())
 
   private fun createMockedRestClient(vararg responses: MockedResponse): RestClient {
     return RestClient(
@@ -51,7 +47,7 @@ internal class CommunityAPIOffenderServiceTest {
   }
 
   private fun offenderServiceFactory(restClient: RestClient): CommunityAPIOffenderService {
-    return CommunityAPIOffenderService(offenderAccessLocation, managedOffendersLocation, staffDetailsLocation, offenderManagersLocation, offenderIdentifiersLocation, responsibleOfficerLocation, restClient, telemetryService)
+    return CommunityAPIOffenderService(offenderAccessLocation, managedOffendersLocation, staffDetailsLocation, offenderManagersLocation, offenderIdentifiersLocation, restClient)
   }
 
   @Test
@@ -96,78 +92,5 @@ internal class CommunityAPIOffenderServiceTest {
     val offenderService = offenderServiceFactory(createMockedRestClient(staffResponse, offendersResponse))
     val result = offenderService.getManagedOffendersForDeliusUser(user)
     assertThat(result).isEqualTo(listOf(Offender("CRN123"), Offender("CRN456")))
-  }
-
-  @Test
-  fun `getResponsibleOfficerDetails returns when there are is a responsible officer`() {
-    val offenderService = offenderServiceFactory(
-      createMockedRestClient(
-        MockedResponse(
-          responsibleOfficerLocation,
-          HttpStatus.OK,
-          "{\n" +
-            "  \"communityOfficer\": {\n" +
-            "    \"code\": \"123\",\n" +
-            "    \"name\": {\n" +
-            "      \"forename\": \"Dan\",\n" +
-            "      \"surname\": \"smith\"\n" +
-            "    },\n" +
-            "    \"username\": \"abcdef\",\n" +
-            "    \"email\": \"dan.smith@gmail.com\",\n" +
-            "    \"responsibleOfficer\": true\n" +
-            "  }\n" +
-            "}",
-        ),
-      ),
-    )
-
-    val responsibleOfficerDetails = offenderService.getResponsibleOfficerDetails("X123456")
-    assertThat(responsibleOfficerDetails?.communityOfficer?.name?.forename).isEqualTo("Dan")
-    assertThat(responsibleOfficerDetails?.communityOfficer?.code).isEqualTo("123")
-    assertThat(responsibleOfficerDetails?.communityOfficer?.responsibleOfficer).isTrue
-  }
-
-  @Test
-  fun `getResponsibleOfficerDetails fails when there are no responsible officers`() {
-    val offenderService = offenderServiceFactory(
-      createMockedRestClient(
-        MockedResponse(
-          responsibleOfficerLocation,
-          HttpStatus.OK,
-          "{\n" +
-            "  \"communityOfficer\": {\n" +
-            "    \"code\": \"123\",\n" +
-            "    \"name\": {\n" +
-            "      \"forename\": \"Dan\",\n" +
-            "      \"surname\": \"smith\"\n" +
-            "    },\n" +
-            "    \"username\": \"abcdef\",\n" +
-            "    \"email\": \"dan.smith@gmail.com\",\n" +
-            "    \"responsibleOfficer\": false\n" +
-            "  }\n" +
-            "}",
-        ),
-      ),
-    )
-
-    assertThrows<InvalidAssumptionError> {
-      offenderService.getResponsibleOfficerDetails("X123456")
-    }
-  }
-
-  @Test
-  fun `getResponsibleOfficerDetails throws not found`() {
-    val offenderService = offenderServiceFactory(createMockedRestClient(MockedResponse(responsibleOfficerLocation, HttpStatus.NOT_FOUND, "")))
-    assertThrows<InvalidAssumptionError> {
-      offenderService.getResponsibleOfficerDetails("X123456")
-    }
-  }
-
-  @Test
-  fun `getResponsibleOfficerDetails throws unhandled error`() {
-    val offenderService = offenderServiceFactory(createMockedRestClient(MockedResponse(responsibleOfficerLocation, HttpStatus.INTERNAL_SERVER_ERROR, "")))
-    assertThrows<WebClientResponseException> {
-      offenderService.getResponsibleOfficerDetails("X123456")
-    }
   }
 }
